requirements.txt
python-dotenv==1.2.1
regex==2024.11.6
schedule==1.2.2
psutil==7.2.2
plexapi==4.18.0
requests==2.32.5
packaging==26.0
ruamel.yaml==0.19.1
colorlog==6.10.1
apprise==1.9.1